refactor(survey-templates): hoist xm-template-ids to shared app/lib
Move xm-template-ids.ts out of the onboarding route folder into
app/lib so it can be shared by the v3 template-create logic without
reaching into onboarding internals. Rename TOnboardingXMTemplateId
to TXMTemplateId to reflect its broader scope and update all imports.
