+++ apps/web/app/api/v3/surveys/templates/lib/template-create.ts
@@ -4,11 +4,11 @@ import { logger } from "@formbricks/logger";
 import { ZSurveyType } from "@formbricks/types/surveys/types";
 import type { TTemplate } from "@formbricks/types/templates";
 import { ZUserLocale } from "@formbricks/types/user";
-import { XM_TEMPLATE_IDS } from "@/app/(app)/(onboarding)/organizations/[organizationId]/workspaces/new/templates/lib/xm-template-ids";
 import { requireV3WorkspaceAccess } from "@/app/api/v3/lib/auth";
 import { problemBadRequest, problemInternalError } from "@/app/api/v3/lib/response";
 import type { TV3AuditLog, TV3Authentication } from "@/app/api/v3/lib/types";
 import { CUSTOM_SURVEY_TEMPLATE_ID, getTemplateById } from "@/app/lib/templates";
+import { XM_TEMPLATE_IDS } from "@/app/lib/xm-template-ids";
 import { replacePresetPlaceholders } from "@/lib/utils/templates";
 import { getWorkspace } from "@/lib/workspace/service";
 import { getTranslate } from "@/lingodotdev/server";
