+++ apps/web/app/lib/xm-template-ids.ts
@@ -1,3 +1,3 @@
 export const XM_TEMPLATE_IDS = ["nps", "star-rating", "csat", "ces", "smileys", "enps"] as const;
 
-export type TOnboardingXMTemplateId = (typeof XM_TEMPLATE_IDS)[number];
+export type TXMTemplateId = (typeof XM_TEMPLATE_IDS)[number];
